+++ meson.build
@@ -102,6 +102,7 @@ sources = [
   'src/game/effect_routines/lara_effects.c',
   'src/game/effect_routines/powerup.c',
   'src/game/effect_routines/raising_block.c',
+  'src/game/effect_routines/sand.c',
   'src/game/effects/blood.c',
   'src/game/effects/exploding_death.c',
   'src/game/effects/gun.c',
@@ -163,7 +164,6 @@ sources = [
   'src/game/objects/effects/gunshot.c',
   'src/game/objects/effects/missile.c',
   'src/game/objects/effects/ricochet.c',
-  'src/game/objects/effects/sand.c',
   'src/game/objects/effects/splash.c',
   'src/game/objects/effects/stairs2slope.c',
   'src/game/objects/effects/turn_180.c',
